refactored starter code
