+++ Starter Application/B07DemoSummer2024/app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:103450168850:android:18be49669e0000df0552cb",
         "android_client_info": {
-          "package_name": "com.example.b07demosummer2024"
+          "package_name": "com.b07group47.taamcollectionmanager"
         }
       },
       "oauth_client": [],
