+++ Starter Application/B07DemoSummer2024/app/src/androidTest/java/com/example/taamcollectionmanager/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.b07demosummer2024;
+package com.example.taamcollectionmanager;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.b07demosummer2024", appContext.getPackageName());
+        assertEquals("com.example.taamcollectionmanager", appContext.getPackageName());
     }
 }
