+++ Starter Application/B07DemoSummer2024/app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <application
+        android:name=".MainActivity"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -10,10 +11,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.B07DemoSummer2024"
+        android:theme="@style/Theme.B07TaamG47"
         tools:targetApi="31">
         <activity
-            android:name=".ItemAdapater"
+            android:name=".ItemAdapter"
             android:exported="false" />
         <activity
             android:name=".Item"
