+++ Starter Application/B07DemoSummer2024/app/src/main/java/com/b07group47/taamcollectionmanager/MainActivity.java
@@ -1,18 +1,10 @@
-package com.example.b07demosummer2024;
+package com.b07group47.taamcollectionmanager;
 
 import android.os.Bundle;
 
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
-
-import android.widget.Button;
-import android.view.View;
-import android.content.Intent;
 
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
