+++ Starter Application/B07DemoSummer2024/app/src/main/java/com/b07group47/taamcollectionmanager/ScrollerFragment.java
@@ -1,10 +1,9 @@
-package com.example.b07demosummer2024;
+package com.b07group47.taamcollectionmanager;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ScrollView;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
