+++ Starter Application/B07DemoSummer2024/app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.B07DemoSummer2024" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.B07TaamG47" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.B07DemoSummer2024" parent="Base.Theme.B07DemoSummer2024" />
+    <style name="Theme.B07TaamG47" parent="Base.Theme.B07TaamG47" />
 </resources>
