+++ Starter Application/B07DemoSummer2024/app/src/test/java/com/example/taamcollectionmanager/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.b07demosummer2024;
+package com.example.taamcollectionmanager;
 
 import org.junit.Test;
